requirements.txt
torch>=1.13
torchvision>=0.14
numpy>=1.22
opencv-python>=4.6
Pillow>=9.0
scipy>=1.9
scikit-learn>=1.1
lmdb>=1.3
pyyaml>=6.0
tqdm>=4.64
tensorboard>=2.11
